Fixed fluxV2 extension resource template (#288)
* Fixed fluxV2 extension resource template

The "configurationSettings" and "configurationProtectedSettings" sections of "Microsoft.KubernetesConfiguration/flux" were defined in the wrong area. Corrected the sections to comply with the schema. https://docs.microsoft.com/en-us/azure/templates/microsoft.kubernetesconfiguration/extensions?tabs=json#template-format
This also fixes the issue of configurations not appearing in the AKS cluster's "GitOps" view in the Azure Portal.

* Fixed case of flux extension type value

Some areas of the resource provider are case sensitive. If using lowercase "microsoft.flux" the GitOps configuration appears in the Azure Portal